backend/requirements.txt
fastapi
pydantic
requests
sqlalchemy
huggingface_hub
python-dotenv
googletrans==4.0.0rc1
uvicorn
pytest
httpx
